Move cohort re-labelling logic to Cohort.ts
