+++ libs/core-ui/src/lib/Cohort/ErrorCohort.ts
@@ -57,49 +57,6 @@ export class ErrorCohort {
     return filtersRelabeled;
   }
 
-  public static getLabeledFilters(
-    filters: IFilter[],
-    jointDataset: JointDataset
-  ): IFilter[] {
-    // return the filters relabeled from Data# to original label
-    const filtersRelabeled = filters
-      .filter((item) => item)
-      .map((filter: IFilter): IFilter => {
-        const label = jointDataset.metaDict[filter.column].label;
-        return {
-          arg: filter.arg,
-          column: label,
-          method: filter.method
-        };
-      });
-    return filtersRelabeled;
-  }
-
-  public static getLabeledCompositeFilters(
-    compositeFilters: ICompositeFilter[],
-    jointDataset: JointDataset
-  ): ICompositeFilter[] {
-    // return the filters relabeled from Data# to original label
-    const filtersRelabeled = compositeFilters.map(
-      (compositeFilter: ICompositeFilter): ICompositeFilter => {
-        if (compositeFilter.method) {
-          return ErrorCohort.getLabeledFilters(
-            [compositeFilter as IFilter],
-            jointDataset
-          )[0] as ICompositeFilter;
-        }
-        return {
-          compositeFilters: ErrorCohort.getLabeledCompositeFilters(
-            compositeFilter.compositeFilters,
-            jointDataset
-          ),
-          operation: compositeFilter.operation
-        } as ICompositeFilter;
-      }
-    );
-    return filtersRelabeled;
-  }
-
   public filtersToString(): string[] {
     const cohortFilters = getBasicFilterString(
       this.cohort.filters,
