+++ libs/core-ui/src/lib/util/calculateBoxData.ts
@@ -2,6 +2,7 @@
 // Licensed under the MIT License.
 
 import { ErrorCohort } from "../Cohort/ErrorCohort";
+import { Cohort } from "../Cohort/Cohort"
 import { IHighchartBoxData } from "../Interfaces/IHighchartBoxData";
 
 export async function calculateBoxPlotDataFromErrorCohort(
@@ -37,12 +38,12 @@ export async function calculateBoxPlotDataFromSDK(
   ) => Promise<IHighchartBoxData>,
   queryClass?: string
 ): Promise<IHighchartBoxData> {
-  const filtersRelabeled = ErrorCohort.getLabeledFilters(
+  const filtersRelabeled = Cohort.getLabeledFilters(
     errorCohort.cohort.filters,
     errorCohort.jointDataset
   );
 
-  const compositeFiltersRelabeled = ErrorCohort.getLabeledCompositeFilters(
+  const compositeFiltersRelabeled = Cohort.getLabeledCompositeFilters(
     errorCohort.cohort.compositeFilters,
     errorCohort.jointDataset
   );
